adding wager payout

+++ Assets/SteamPlayer.cs
@@ -5,7 +5,7 @@ using Steamworks;
 public class SteamPlayer : NetworkBehaviour
 {
     public string playerName;
-    public NetworkVariable<float> credits;
+    public NetworkVariable<int> credits;
 
     public void Start()
     {
